requirements.txt
tavily-python
langgraph
langchain-core
openai
python-dotenv
pydantic
typing_extensions
numpy          
tensorflow 
transformers
tf-keras